refactor(gsd-db): peel worktree reconciliation into db/writers/reconcile.ts
Phase A, commit 3. Extracts copyWorktreeDb + reconcileWorktreeDb (487 lines —
the largest non-core subsystem) into db/writers/reconcile.ts.

Lifts the shared TERMINAL_STATUS_SQL fragment into db/sql-constants.ts (imported
as a value by reconcile and the core slice/task writers, not re-exported, so it
stays an implementation detail). Drops the now-unused openDatabase and node:fs
imports from gsd-db.ts (reconcile was their only internal user).

gsd-db.ts: 2,644 → 2,158 lines.

Verified: typecheck:extensions clean; single-writer-invariant + full
reconcileWorktreeDb merge/no-downgrade/seq-preservation suite + merge-db-cycle
green (49 tests).
